backend/requirements.txt
flask
scikit-learn
PyPDF2
python-docx
numpy